+++ workers/tests/integrationtests/test_base_indicator.py
@@ -3,9 +3,7 @@ import unittest
 from unittest import mock
 
 from ohsome_quality_analyst.geodatabase import client as db_client
-from ohsome_quality_analyst.indicators.ghs_pop_comparison_buildings.indicator import (
-    GhsPopComparisonBuildings,
-)
+from ohsome_quality_analyst.indicators.minimal.indicator import Minimal
 
 
 class TestBaseIndicator(unittest.TestCase):
@@ -13,10 +11,10 @@ class TestBaseIndicator(unittest.TestCase):
         self.feature = asyncio.run(
             db_client.get_feature_from_db(dataset="regions", feature_id="3")
         )
-        self.layer_name = "building_count"
+        self.layer_name = "minimal"
 
     def test_as_feature(self):
-        indicator = GhsPopComparisonBuildings(feature=self.feature, layer=mock.Mock())
+        indicator = Minimal(feature=self.feature, layer=mock.Mock())
 
         feature = indicator.as_feature()
         assert feature.is_valid
@@ -26,23 +24,16 @@ class TestBaseIndicator(unittest.TestCase):
         assert "data" not in feature["properties"]
 
     def test_as_feature_include_data(self):
-        indicator = GhsPopComparisonBuildings(feature=self.feature, layer=mock.Mock())
+        indicator = Minimal(feature=self.feature, layer=mock.Mock())
 
         feature = indicator.as_feature(include_data=True)
         assert feature.is_valid
         for key in ("result", "metadata", "layer", "data"):
             assert key in feature["properties"]
-        for key in (
-            "pop_count",
-            "area",
-            "pop_count_per_sqkm",
-            "feature_count",
-            "feature_count_per_sqkm",
-        ):
-            assert key in feature["properties"]["data"]
+        assert "count" in feature["properties"]["data"]
 
     def test_as_feature_flatten(self):
-        indicator = GhsPopComparisonBuildings(feature=self.feature, layer=mock.Mock())
+        indicator = Minimal(feature=self.feature, layer=mock.Mock())
         feature = indicator.as_feature(flatten=True)
         assert feature.is_valid
         for key in (
@@ -53,11 +44,11 @@ class TestBaseIndicator(unittest.TestCase):
             assert key in feature["properties"]
 
     def test_data_property(self):
-        indicator = GhsPopComparisonBuildings(feature=self.feature, layer=mock.Mock())
+        indicator = Minimal(feature=self.feature, layer=mock.Mock())
         self.assertIsNotNone(indicator.data)
         for key in indicator.data.keys():
             self.assertNotIn(key, ("result", "metadata", "layer", "feature"))
 
     def test_attribution_class_property(self):
-        self.assertIsNotNone(GhsPopComparisonBuildings.attribution())
-        self.assertIsInstance(GhsPopComparisonBuildings.attribution(), str)
+        self.assertIsNotNone(Minimal.attribution())
+        self.assertIsInstance(Minimal.attribution(), str)