+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 - Remove unused report `RemoteMappingLevelOne` ([#380])
 - Substitute Simple Report with a Report named Minimal for testing purposes ([#342] [#385])
+- Add a minimal Indicator for testing purposes ([#383])
 
 ### How to Upgrade
 
@@ -32,6 +33,7 @@
 [#375]: https://github.com/GIScience/ohsome-quality-analyst/pull/375
 [#380]: https://github.com/GIScience/ohsome-quality-analyst/pull/380
 [#382]: https://github.com/GIScience/ohsome-quality-analyst/pull/382
+[#383]: https://github.com/GIScience/ohsome-quality-analyst/pull/383
 [#385]: https://github.com/GIScience/ohsome-quality-analyst/pull/385